Remove all gptapi code and associated files
This commit removes all the code and associated files related to the gptapi feature. In addition to various configuration and property files, model, controller and client code files have also been removed. It also includes the removal of test setup files, ensuring a clean removal of all functionalities related to this feature.